Create Prims_MST_data.txt
This input file describes an undirected graph with integer edge costs. It has the format
[number_of_nodes] [number_of_edges]
[one_node_of_edge_1] [other_node_of_edge_1] [edge_1_cost]
[one_node_of_edge_2] [other_node_of_edge_2] [edge_2_cost]
...
For example, the third line of the file is "2 3 -8874", indicating that there is an edge 
connecting vertex #2 and vertex #3 that has cost -8874. The below algorithm assumes that the 
graph has no parallel edges. No assumption are made about the edge costs being positive, or 
being distinct.